news

+++ Routes/taskRoutes.js
@@ -1,29 +1,33 @@
 const express = require("express");
 const router = express.Router();
-const Task = require("../Models/Task.js");
+const Task = require("../Models/Task");
+const authMiddleware = require("../Middleware/authMiddleware");
 
-router.get("/", async (req, res) => {
+router.get("/", authMiddleware, async (req, res) => {
   try {
-    const tasks = await Task.find();
+    const tasks = await Task.find({ user: req.user.userId });
     res.status(200).json(tasks);
   } catch (error) {
     res.status(500).json({ message: error.message });
   }
 });
 
-router.post("/", async (req, res) => {
+router.post("/", authMiddleware, async (req, res) => {
   try {
-    const task = await Task.create(req.body);
+    const task = await Task.create({ ...req.body, user: req.user.userId });
     res.status(201).json(task);
   } catch (error) {
     res.status(400).json({ message: error.message });
   }
 });
 
-router.put("/:id", async (req, res) => {
+router.put("/:id", authMiddleware, async (req, res) => {
   try {
     const task = await Task.findByIdAndUpdate(
-      req.params.id,
+    {
+      _id: req.params.id,
+      user: req.user.userId,
+    },
       req.body,
       { new: true }
     );
@@ -34,9 +38,9 @@ router.put("/:id", async (req, res) => {
   }
 });
 
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", authMiddleware, async (req, res) => {
   try {
-    await Task.findByIdAndDelete(req.params.id);
+    await Task.findByIdAndDelete({ _id: req.params.id, user: req.user.userId });
 
     res.status(200).json({
       message: "Task deleted",
